add check on common redhat prerequisites script in current directory

+++ gluex_prereqs_centos_7.sh
@@ -1,4 +1,10 @@
 #!/bin/sh
 yum -y install epel-release
-./prereqs_redhat.sh
+file=prereqs_redhat.sh
+if [ ! -f $file ]
+    then
+    echo $file not found in current directory
+    exit 1
+fi
+./$file
 yum -y install lapack-static blas-static
